Making Progress on dynamic changing of the mud line and water line.
Added the EKS blender file to the repo (the blend1 file is the last saved version as a way to recover changes)

Updated the blender object so all the rods are unique and added inner corner points.

Updated the waste material and temporarily are calling it oil.mat.

Updated our prefabs to be mostly correct and deleted the old ones.

Added scripts for simple camera control and filling the whole trapezoid with the oil (more to come on this later).

Updated the project version