Great first start on the customizable water and oil levels.
ToDo: Water level can go down and each rod has it's own control area.

Prefab - I updated the prefab so the parent is the controller of the children.

DemoScene - Deactivated the container for transparency issues.

OilController - This one is hefty, the rest of the comments cover it... Sorry

After 2 days of figuring out the math, we are doing an inverse of the Slope-Intercept to find the slope and intercept of each angle along the container.

Start - Assign our meshes and go ahead and calculate all of our mathy stuff. Finally go ahead and assign the vertices (corners) of our cubes so there isn't a weird flash.

Update - We are updating the oil-water line with user input to demonstrate the new capibility. And then reassign our verts based on the new oil-water line.

Calculate Slopes - As I mentioned before this is an inverse take on the Slope-Intercept form, so instead of rise over run, it is run over rise and forward over rise. (Makes the math a bit easier since we always know the y)

Get...AtWaterLine - Uses our Slope Interecept form to figure out the x and z of the points.

GetVerticesFor... - Correctly orders and returns the points based on our position variables.