Huge Refactor Of The Model Controller (will rename some stuff like "oil" later)
Added a variable for the top of the water.

Condensed 24 variables into 6. :)

Moved the slope and intercept calculations into methods

Made the functions that get the vertices generic, so you give it a y and it tells you the coordinates.

We now control both the top and bottom of both the water and oil.